backend/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6
python-dotenv>=1.0.0

# Document Processing
pypdf2>=3.0.1
python-docx>=1.1.0

# LLM - Google Gemini
google-generativeai>=0.3.1,<0.4.0
langchain>=0.1.5
langchain-google-genai>=0.0.6

# Database
neo4j>=5.17.0
chromadb>=0.4.22

# Utilities
pydantic>=2.0.0
httpx>=0.26.0
sentence-transformers>=2.2.0
groq>=0.4.0
